# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /transactionservice/transaction/{transaction_id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    put:
      description: Insert transaction with an id
      # used as the method name of the controller
      operationId: addTxn
      parameters:
        - name: transaction_id
          in: path
          description: ID of the txn to be inserted
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Get transaction details mapped with the id
      # used as the method name of the controller
      operationId: getTxn
      parameters:
        - name: transaction_id
          in: path
          description: ID of the txn to be fetched
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetTxnResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /transactionservice/types/{type}:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Get all txn for the specified type
      # used as the method name of the controller
      operationId: getTxnByType
      parameters:
        - name: type
          in: path
          description: Filter by type of txn
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetTypeResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /transactionservice/sum/{transaction_id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns sum of all txn down the chain
      # used as the method name of the controller
      operationId: getTxnSum
      parameters:
        - name: transaction_id
          in: path
          description: Txn id for which sum has to be computed
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetSumResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - status
    properties:
      status:
        type: string
  GetTxnResp:
    required:
      - amount
      - type
    properties:
      amount:
        type: number
      type:
        type: string
  GetSumResp:
    required:
      - sum
    properties:
      sum:
        type: number
  GetTypeResp:
    properties:
      txn_list:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string